* baagaard/fix-compilation-warnings:
  Code cleanup to remove compilation warnings.
  Cleanup to eliminate compilation warnings.
  Cleanup to fix compilation warnings.
  Cleanup to fix compilation warnings.
  Cleanup cover thumbnail.

+++ libsrc/pylith/feassemble/GeometryHex3D.cc
@@ -184,7 +184,7 @@ pylith::feassemble::GeometryHex3D::jacobian(scalar_array* jacobian,
 	 this->numCorners()+1 == numVertices); // quadratic hex
   assert(this->spaceDim() == spaceDim);
   assert(this->cellDim() == cellDim);
-  assert(spaceDim*cellDim == jacobian->size());
+  assert(size_t(spaceDim*cellDim) == jacobian->size());
 
   const PylithScalar x0 = vertices[0];
   const PylithScalar y0 = vertices[1];
